Moved JSON reader to io
